Make additional formatting changes to 'Embedding Null Code'
This change addresses the feedback from pull request #79, and also adds the
author & contributor data to the header of the file. A pair of script tags
from the examples of the initial page were also re-added, as their formatting
on the contemporary page made them look like a mistaken artifact of the
conversion.

Original source:
https://wiki.owasp.org/index.php/Embedding_Null_Code