Fix small typos and adjust spacing in updated FAQ sections
I missed a few things during my previous passes, and also noticed that
the spacing of the initial questions' response could be adjusted. It
should be noted that line breaks in markdown do not directly translate
to paragraph breaks for the end reader, who I prioritise the readability
of text for.