Close #120, broken links for application testing (#131)
* Add sample application links to FAQ, update metadata and first two questions

There was a query about the broken links under the "Where can I try out
my testing skills?" heading; I began to start updating the page as a
whole, but realised this is going to be a project all in itself.

* Fix small typos and adjust spacing in updated FAQ sections

I missed a few things during my previous passes, and also noticed that
the spacing of the initial questions' response could be adjusted. It
should be noted that line breaks in markdown do not directly translate
to paragraph breaks for the end reader, who I prioritise the readability
of text for.