Removed duplicate entry for RIPS
RIPS is included twice, the latter record indicates that the tool is
open source or free.  However, I cannot find the claim of being open
source on the current website, and the repository on github
(https://github.com/ripsscanner/rips) has not been updated for 6 years.
Keep the older record that marks the tool as commercial seems to make
sense.