Update _data/tools.json for RIPS (#310)
* Removed abandoned RIPS

The website itself says "... development is abandoned since 2013 due to its fundamental limitations.". Also, the new RIPS project is already included, so it makes sense to remove this old entry

* Removed duplicate entry for RIPS

RIPS is included twice, the latter record indicates that the tool is
open source or free.  However, I cannot find the claim of being open
source on the current website, and the repository on github
(https://github.com/ripsscanner/rips) has not been updated for 6 years.
Keep the older record that marks the tool as commercial seems to make
sense.