Add more up to date examples for .NET (#365)
* Add more up to date examples for .NET

This commit adds more up to date examples of performing certificate pinning in .NET. Using `ServicePointManager.ServerCertificateValidationCallback` affects all requests from the AppDomain, so care must be taken to handle server validation for requests to domains other than the one to which we want to apply certificate pinning. In more recent versions of .NET Framework, server certificate validation can be performed per request, offering better control over `ServicePointManager` on when to apply validation logic.

Usage of `WebRequest` and `HttpWebRequest` have long been superseded by `HttpClient`, which exposes the ability to specify a server certificate validation callback on the `HttpClientHandler` used by the client. This commit adds an example.

* Update examples

This commit updates examples following PR review #365.
When the request is to the target domain for which we want to
perform certificate pinning, pinning will always be performed.
For any other domain, validity is determined based on whether
there are any policy errors.

* tildes, be gone!