Add some links to abbreviations to introduce them
Problem is that some abreviations are used prior they are
explained in detail in further sections. As novice I have no
glue what STRIDE or DREAD is until I've read the full article.

For beginners in the field it is helpful to just simply
click to the according wikipedia article to see what STRIDE,
DREAD etc. instead of googling it.

This adds links to Wikipedia for all abreviations I spotted
in the text.

Signed-off-by: Sven Strittmatter <ich@weltraumschaf.de>