No leading whitespace in tools title (#807)
* Modify tools schema to enforce no leading leading whitespace in title

The tools are sorted by title when constructing the HTML table.
Leading whitespace causes an unintended sort order.  Add a pattern
for the title property to enforce no leading whitespace.

* Run "Validate Tools JSON" workflow when tools schema is modified

The workflow was only triggered to run when `_data/tools.json` was
modified.  It should also run if `tools_schema.json` is modified.

* Remove leading whitespace from tool titles