Create Server-Side Request Forgery (SSRF)
Server-Side Request Forgery (SSRF) is a type of exploit where an attacker abuses functionality on the server to read or update internal resources. The attacker can supply or modify a URL to which the code running on the server will read or submit data. By carefully selecting the URLs, the attacker may be able to read server configuration such as AWS metadata, connect to internal services like HTTP-enabled databases, or perform post requests towards internal services that are not intended to be exposed.