Create RSQL_Injection.md (#1086)
* Create RSQL_Injection.md

RSQL Injection is a vulnerability in web applications that use RSQL as a query language in RESTful APIs. Similar to SQL Injection and LDAP Injection, this vulnerability occurs when RSQL filters are not properly sanitized, allowing an attacker to inject malicious queries to access, modify or delete data without authorization.

---------

Co-authored-by: Rick M <kingthorin@users.noreply.github.com>